Improve timing and add OKTO live drum kits
Preserve fractional piano-roll timing through MIDI recording, playback,
editing, takes, state and mode conversion. Export both rolls at 9600 PPQ
and switch empty projects between modes without a confirmation.

Add seven percussion sounds and ten empty Live Drums kits. Match the
OKTO B physical pad order, including kick on channel 5 and snare on 6.
Preserve saved custom input assignments.

Keep Duck and OV Off effective on tails from previous live patterns,
verify Choke on/off, and clarify the Overlap tooltip. Retain version 1.5.8.

Validation: all 22 offline regression programs pass; LiveDrumTest UI and
live dynamics checks pass on arm64 and x86_64. All 274 sounds pass UIAudit;
C4 and native-tuning sound comparisons are documented in docs/audits.

+++ tests/GenTest.cpp
@@ -1657,7 +1657,7 @@ int main()
         {
             const auto& dn = C4c.drawNotes[i];
             if (dn.start != 0)   // bar anchors are licensed to overlap
-            { if (melOcc2[juce::jlimit(0, 15, dn.start / 24)]) ++under; else ++inGap; }
+            { if (melOcc2[juce::jlimit(0, 15, (int)(dn.start / 24))]) ++under; else ++inGap; }
             if (dn.semi > melMed2) ++laneViol;
         }
         printf("  SCORE arrangement comp           inGap=%d under=%d total=%d melMed=%d\n",
